@@ -12,8 +12,10 @@
     "eslint-loader": "4.0.2",
     "mini-css-extract-plugin": "2.2.0",
     "node-sass": "6.0.1",
-    "sass-loader": "12.1.0",
-    "webpack": "5.50.0",
+    "sass": "^1.49.8",
+    "sass-loader": "^12.6.0",
+    "svgo-loader": "^3.0.0",
+    "webpack": "^5.69.1",
     "webpack-cli": "4.8.0"
   },
   "scripts": {
@@ -21,6 +23,10 @@
     "watch": "webpack --watch"
   },
   "dependencies": {
-    "vue": "^3.2.31"
+    "axios": "^0.26.0",
+    "file-loader": "^6.2.0",
+    "intl": "^1.2.5",
+    "vue": "^3.2.31",
+    "vue-loader": "^17.0.0"
   }
 }
