@@ -1,7 +1,7 @@
 module.exports = {
     "extends": "eslint:recommended",
     "parserOptions": {
-      "ecmaVersion": 7,
-      "sourceType": "module"
+        "sourceType": "module",
+        "ecmaVersion": 2020
     }
 }
